Language.Copilot.Compiler

Description

Transform the copilot specification in an atom one, and then compile that one.

Documentation

copilotToAtom :: StreamableMaps Spec -> Sends -> Maybe Period -> [(Var, String)] -> (Period, Atom ())

Compiles an Copilot specification to an Atom one. The period is given as a Maybe : if it is Nothing, an optimal period will be chosen.

tmpSampleStr :: String