Language.Copilot.Interface

Description

Used by the end-user to easily give its arguments to dispatch.

Documentation

data Options

baseOpts :: Options

test :: Int -> Options -> IO ()

interpret :: Streams -> Int -> Options -> IO ()

compile :: Streams -> Name -> Options -> IO ()

verify :: FilePath -> Int -> IO ()

interface :: Options -> IO ()

The main function that dispatches.

help :: IO ()

setE :: Vars -> Options -> Options

Sets the environment for simulation by giving a mapping of external variables to lists of values. E.g.,

 setE (emptySM {w32Map = fromList [("ext", [0,1..])]}) ...

sets the external variable names ext to take the natural numbers, up to the limit of Word32.

setC :: String -> Options -> Options

Sets the options for the compiler, e.g.,

 setC "-O2" ...

Sets gcc options.

setO :: Name -> Options -> Options

setP :: Period -> Options -> Options

Manually set the period for the program. Otherwise, the minimum required period is computed automatically. E.g.,

setP 100 ... sets the period to be 100. The period must be at least 1.

setI :: Options -> Options

setPP :: (String, String) -> Options -> Options

Sets the code to precede and follow the copilot specification If nothing, then code appropriate for communication with the interpreter is generated

setN :: Int -> Options -> Options

Sets the number of iterations of the program to simulation:

setN 50 simulations the program for 50 steps.

setV :: Verbose -> Options -> Options

Set the verbosity level.

setR :: Int -> Options -> Options

setDir :: String -> Options -> Options

Sets the directory into which the output of compiled programs should be placed. If the directory does not exist, it will be created.

setGCC :: String -> Options -> Options

Sets the compiler to use, given as a path to the executable. The default is "gcc".

setArrs :: [(String, Int)] -> Options -> Options

Set the external arrays.

setClock :: Clock -> Options -> Options

Sets the hardware clock. See http:github.comtomahawkinsatomblobmasterLanguageAtom/Code.hs.

module Language.Copilot.Dispatch