Language.Copilot.Interpreter

Description

The Copilot interpreter.

Documentation

interpretStreams :: StreamableMaps Spec -> Vars -> Vars

The main function of this module. It takes a Copilot specification, the values of the monitored values, and returns the values of the streams.