Language.Copilot.Language.RandomOps

Description

Sets of operators for Tests.Random.hs

Documentation

opsF :: Operators

opsF2 :: Operators

opsF, opsF2 and opsF3 are fed to Tests.Random.randomStreams. They allows the random generated streams to include lots of operators. If you add a new operator to Copilot, it would be nice to add it to one of those, that way it could be used in the random streams used for testing. opsF holds all the operators of arity 1, opsF2 of arity 2 and opsF3 of arity3 They are StreamableMaps, because operators are sorted based on their return type.

opsF3 :: Operators

module Language.Copilot.Tests.Random