Language.Copilot.Language.Sampling

Description

Functions for sampling external variables.

The next functions provide easier access to typed external variables.

extB :: ExtCl a => a -> Spec Bool

extI8 :: ExtCl a => a -> Spec Int8

extI16 :: ExtCl a => a -> Spec Int16

extI32 :: ExtCl a => a -> Spec Int32

extI64 :: ExtCl a => a -> Spec Int64

extW8 :: ExtCl a => a -> Spec Word8

extW16 :: ExtCl a => a -> Spec Word16

extW32 :: ExtCl a => a -> Spec Word32

extW64 :: ExtCl a => a -> Spec Word64

extF :: ExtCl a => a -> Spec Float

extD :: ExtCl a => a -> Spec Double

The next functions provide easier access to typed external arrays.

extArrB :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Bool

extArrI8 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Int8

extArrI16 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Int16

extArrI32 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Int32

extArrI64 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Int64

extArrW8 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Word8

extArrW16 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Word16

extArrW32 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Word32

extArrW64 :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Word64

extArrF :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Float

extArrD :: (ExtCl a, Streamable b, IntegralE b) => a -> Spec b -> Spec Double