Language.Copilot.Libs.Statistics

Description

Basic bounded statistics. In the following, a bound n is given stating the number of periods over which to compute the statistic (n == 1 computes it only over the current period).

Documentation

max :: (Streamable a, NumE a) => Int -> Spec a -> Spec a

Maximum value.

min :: (Streamable a, NumE a) => Int -> Spec a -> Spec a

Minimum value.

sum :: (Streamable a, NumE a) => Int -> Spec a -> Spec a

Summation.

mean :: (Streamable a, Fractional a, NumE a) => Int -> Spec a -> Spec a

Mean value. n must not overflow for word size a for streams over which computation is peformed.

meanNow :: (Streamable a, IntegralE a) => [Spec a] -> Spec a

Mean value over the current set of specs passed in.