Language.Copilot.Language

Description

Describes the language Copilot.

If you wish to add a new operator, the only modification needed is adding it in this module. But if you want it to be used in the random generated streams, add it to either opsF, opsF2 or opsF3

Operators and functions

mod :: (Streamable a, IntegralE a) => Spec a -> Spec a -> Spec a

div :: (Streamable a, IntegralE a) => Spec a -> Spec a -> Spec a

Beware : crash without any possible recovery if a division by 0 happens. Same risk with mod. Use div0 and mod0 if unsure.

mod0 :: (Streamable a, IntegralE a) => a -> Spec a -> Spec a -> Spec a

div0 :: (Streamable a, IntegralE a) => a -> Spec a -> Spec a -> Spec a

As mod and div, except that if the division would be by 0, the first argument is used as a default.

(<) :: (Streamable a, OrdE a) => Spec a -> Spec a -> Spec Bool

(<=) :: (Streamable a, OrdE a) => Spec a -> Spec a -> Spec Bool

(==) :: (Streamable a, EqE a) => Spec a -> Spec a -> Spec Bool

(/=) :: (Streamable a, EqE a) => Spec a -> Spec a -> Spec Bool

(>=) :: (Streamable a, OrdE a) => Spec a -> Spec a -> Spec Bool

(>) :: (Streamable a, OrdE a) => Spec a -> Spec a -> Spec Bool

not :: Spec Bool -> Spec Bool

(||) :: Spec Bool -> Spec Bool -> Spec Bool

(&&) :: Spec Bool -> Spec Bool -> Spec Bool

(^) :: Spec Bool -> Spec Bool -> Spec Bool

(==>) :: Spec Bool -> Spec Bool -> Spec Bool

Boolean constants

data Bool

Constructors

False

True

Arithmetic operators (derived)

class (Eq a, Show a) => Num a where

Basic numeric class.

Minimal complete definition: all except negate or (-)

Methods

(+) :: a -> a -> a

(*) :: a -> a -> a

(-) :: a -> a -> a

negate :: a -> a

Unary negation.

abs :: a -> a

Absolute value.

signum :: a -> a

Sign of a number. The functions abs and signum should satisfy the law:

 abs x * signum x == x

For real numbers, the signum is either -1 (negative), 0 (zero) or 1 (positive).

fromInteger :: Integer -> a

Conversion from an Integer. An integer literal represents the application of the function fromInteger to the appropriate value of type Integer, so such literals have type (Num a) => a.

Division

class Num a => Fractional a where

Fractional numbers, supporting real division.

Minimal complete definition: fromRational and (recip or (/))

Methods

(/) :: a -> a -> a

fractional division

mux :: Streamable a => Spec Bool -> Spec a -> Spec a -> Spec a

Beware : both sides are executed, even if the result of one is later discarded

Copilot variable declarations.

var :: Streamable a => Var -> Spec a

Useful for writing libraries.

varB :: Var -> Spec Bool

varI8 :: Var -> Spec Int8

varI16 :: Var -> Spec Int16

varI32 :: Var -> Spec Int32

varI64 :: Var -> Spec Int64

varW8 :: Var -> Spec Word8

varW16 :: Var -> Spec Word16

varW32 :: Var -> Spec Word32

varW64 :: Var -> Spec Word64

varF :: Var -> Spec Float

varD :: Var -> Spec Double

Copilot constant declarations. For the most part, these are

const :: Streamable a => a -> Spec a

Coerces a type that is Streamable into a Copilot constant.

constI8 :: Int8 -> Spec Int8

constI16 :: Int16 -> Spec Int16

constI32 :: Int32 -> Spec Int32

constI64 :: Int64 -> Spec Int64

constW8 :: Word8 -> Spec Word8

constW16 :: Word16 -> Spec Word16

constW32 :: Word32 -> Spec Word32

constW64 :: Word64 -> Spec Word64

constF :: Float -> Spec Float

constD :: Double -> Spec Double

Boolean stream constants

true :: Spec Bool

false :: Spec Bool

module Language.Copilot.Language.Sampling

Constructs of the copilot language

drop :: Streamable a => Int -> Spec a -> Spec a

Drop i elements from a stream.

(++) :: Streamable a => [a] -> Spec a -> Spec a

Just a trivial wrapper over the Append constructor

(.=) :: Streamable a => Spec a -> Spec a -> Streams

Define a stream variable.

The next functions help typing the send operations

send :: Streamable a => String -> Port -> Spec a -> Streams

Takes a function name, a port number, and a Copilot variable v and constructs a call to the C function name(x,y) where x is the value of the Copilot stream v and y is the port number.

port :: Int -> Port

Triggers

module Language.Copilot.Language.FunctionCalls

Safe casting

module Language.Copilot.Language.Casting

notConstVarErr :: Streamable a => Spec a -> (ArgConstVar -> b) -> b

If the Spec isn't a Var or Const, then throw an error; otherwise, apply the function.